+++ src/io/github/mstehula/elements/Sand.java
@@ -1,6 +1,6 @@
-package io.github.mstehula.element;
+package io.github.mstehula.elements;
 
-import io.github.mstehula.element.interfaces.Moveable;
+import io.github.mstehula.elements.interfaces.Moveable;
 
 /**
  * Created by MStehula on 7/6/2015.
