+++ src/io/github/mstehula/elements/Water.java
@@ -1,8 +1,8 @@
-package io.github.mstehula.element;
+package io.github.mstehula.elements;
 
-import io.github.mstehula.element.interfaces.Liquid;
-import io.github.mstehula.element.interfaces.Moveable;
-import io.github.mstehula.element.interfaces.Solvent;
+import io.github.mstehula.elements.interfaces.Liquid;
+import io.github.mstehula.elements.interfaces.Moveable;
+import io.github.mstehula.elements.interfaces.Solvent;
 
 /**
  * Created by MStehula on 7/6/2015.
