+++ src/io/github/mstehula/elements/interfaces/Liquid.java
@@ -1,4 +1,4 @@
-package io.github.mstehula.element.interfaces;
+package io.github.mstehula.elements.interfaces;
 
 /**
  * Created by MStehula on 7/6/2015.
